requirements.txt
aniso8601==1.2.0
appnope==0.1.0
APScheduler==3.3.1
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
cachetools==2.0.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
colorlog==3.0.1
constantly==15.1.0
cryptography==2.0.3
cssselect==1.0.1
decorator==4.1.2
Flask==0.12.1
Flask-BasicAuth==0.2.0
Flask-RESTful==0.3.5
flask-restful-swagger==0.19
Flask-SQLAlchemy==2.2
frontera==0.7.1
happybase==1.1.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
influxdb==4.1.1
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
kafka-python==1.3.4
lxml==3.8.0
MarkupSafe==1.0
msgpack-python==0.4.8
mysql-connector-python-rf==2.2.2
parsel==1.2.0
pexpect==4.2.1
pickleshare==0.7.4
ply==3.10
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycparser==2.18
PyDispatcher==2.0.5
pydot==1.2.3
Pygments==2.2.0
pymongo==3.5.1
PyMySQL==0.7.11
pyOpenSSL==17.2.0
pyparsing==2.2.0
pytest==3.2.2
python-dateutil==2.6.0
python-json-logger==0.1.8
pytz==2017.2
pyzmq==16.0.2
queuelib==1.4.2
redis==2.10.6
requests==2.13.0
requests-file==1.4.2
Scrapy==1.4.0
scrapy-crawlera==1.2.4
scrapy-splash==0.7.2
scrapyd==1.2.0a1
scrapyd-client==1.2.0a1
selenium==3.6.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.10.0
SpiderKeeper==1.2.0
SQLAlchemy==1.1.9
thriftpy==0.3.9
tldextract==2.1.0
traitlets==4.3.2
Twisted==17.5.0
tzlocal==1.3
urllib3==1.22
w3lib==1.18.0
wcwidth==0.1.7
Werkzeug==0.12.1
zope.interface==4.4.2
